Initial attempt at using cURL.
This is probably just going to be too much of a pain in the arse -- it's
going to be complex to handle the CONNECT request, and even harder
dealing with the TPM -- we don't know what SSL engine cURL will be
compiled with. And we have to be linked against OpenSSL _anyway_.

The only thing cURL buys us, really, is the redirect handling and HTTP
response parsing -- and most of that is trivial enough anyway.

Left here just in case I change my mind when it comes to actually
submitting the forms...