Work around OpenSSL bug with certificate chains.
This will probably be RT#1942 -- OpenSSL will look up issuer
certificates by name, but there might be more than one certificate in
the trust chain with the same name, and it doesn't make sure it gets the
right one. The server suffers this bug too, which is why the client has
to submit the full trust chain with its own certificate.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>