Create libopenconnect.a for GUI authentication dialogs to use.
Now that things have stabilised, it ought to be feasible for us to put
the NetworkManager auth-dialog in the network-manager-openconnect
package where it belongs. Knetworkmanager support for openconnect will need
to use it too.

A static library is the first step; ideally we'll be able to do a sane
dynamic library with a reasonable stable ABI and no namespace pollution.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>