Add 'reconnect' invocation of vpnc-script, to re-ensure routing/DNS setup
If we reconnect because of an actual local network disconnect/reconnect, then
something (DHCP, etc.) may have screwed up the routing and DNS according to
the local configuration. Give the script a chance to remedy that.

With iproute (i.e. modern Linux) it ought to work just to make vpnc-script
do the same as it does on 'connect'. For other systems it's somewhat harder.

For now vpnc-script will ignore it, anyway.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>