Add openconnect_vpninfo_new_with_cbdata() function to ease C++ integration
C++ callers really want the 'self' object pointer to be the first argument
of the callbacks. Give them the chance to get that, instead of the vpninfo
pointer.

Preserve the old openconnect_vpninfo_new() call, even with the same
prototype for the callback functions, for compatibility with the existing
GNOME auth-dialog.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>