Use TLSv1 again, but with no extensions.
In commit 3bee59cd8c4fd6dc38bc2c7d5edb9b5795509fca ("Use SSLv3 not TLSv1")
we switched to SSLv3 to avoid problems with VPN servers (or firewalls)
which reject extensions in ClientHello. Another user now reports that
using SSLv3 is failing, and he needs to use TLSv1.

Testing confirms that the originally problematic server *does* work with
TLSv1, as long as we disable the session ticket extension. So that makes
everyone happy... for now, until a new extension is invented and enabled
by default, and we have to block that too.  It's a shame that there's no
SSL_OP_NO_EXTENSIONS which would turn them *all* off.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>