Use I_LINK instead of I_PLINK on Solaris
This should avoid leaving stale tunnel devices around, because they'll
automatically go away when the fd gets closed.

This is only for Legacy IP though; We are still plumbing for IPv6 in the
vpnc-script, which is weird and inconsistent. If only I could work out
how to do it in tun.c like we do for IPv4...

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>