Invoke vpnc-script with "pre-init" to load tun module if necessary
Shift the os-specific tun setup out into a separate function, just because
it declares local variables and the pre-init script call is now happening
before that in the code block.

Clean up script invocation a little too, and fix the fact that we'd run the
script with "reconnect" reason even if there wasn't a script at all.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>