Attempt to build against OpenSSL without pkg-config
If we insist on using --with-openssl to do that, it looks for (and
builds) a static library. That's intended for use with a locally-built
OpenSSL source tree. For systems like FreeBSD which lack pkg-config for
an otherwise normally-installed dynamic OpenSSL library, we need to cope
the same way as we do with pkg-config-less zlib.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>