Clean up CSD invocation for XML POST
We don't use a CSD trojan to download; we *require* a local wrapper.

Theoretically we ought to be able to invoke a 'real' Cisco hostscan
tool. We ought to fix the command line arguments for that but let's keep
it simple for now. Just keep the command line exactly the same as for
wrapping the trojan, with an empty string where the name of the
downloaded file would be.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>