Close HTTPS socket after various errors
This avoids leaving the socket in an unknown state. We were attempting to
send a request with a stale or out-of-sync socket, and that would make
the *next* request fail too, when it should have opened a new connection
for itself.

We should also make do_https_request() notice that and actually retry for
itself when it fails to even *send* the request, if it was re-using an
already open socket. But currently it doesn't *know* if it's re-using a
socket so that'll require a little more work.

Fixes Debian bug #708928: http://bugs.debian.org/708928

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>