Include string.h where needed
This fixes the Solaris build; reported by "Rodney".

Also clean up #includes in compat.c; there's no need for them to be
hidden within the various ifdef blocks. Leave <sys/time.h> where it was,
just in case it's problematic on some platform. The rest are already
included unconditionally in other files, so should be fine.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>