library: Add openconnect_set_xmlpost() to support different CSD modes
In certain configurations, XML POST mode can impact the use of
CSD/Hostscan.  Some gateways are set up to refuse the CSD results if
XML POST is not being used; but in other cases supporting XML POST
on the client side can require a custom CSD wrapper script which isn't
needed for the non-XML-POST case.  So we'll allow the caller to decide
which mode to use.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>