Include supporting certificates from PKCS#11 tokens
Sometimes, we really need to include the supporting CA on the wire, and
it is only found in the PKCS#11 alongside the user's own certificate.

Seen with the Belgian eID devices, and a server which didn't appear to
know the 'Citizen CA' for itself, but did trust 'Belguim Root CA2'.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>