dtls: Fix dtls_pkt heap overflow (CVE-2013-7098)
Each time openconnect (re)establishes a CSTP connection, it reinitializes
vpninfo->actual_mtu from the X-{CSTP,DTLS}-MTU headers provided by the
gateway.  dtls_mainloop() uses the CURRENT value of vpninfo->actual_mtu
as an upper limit when reading a packet, but it used the ORIGINAL value
of vpninfo->actual_mtu when allocating the dtls_pkt buffer.  So if the
original value was smaller than the current value, it is possible to
overwrite a part of the heap with packet data.

This probably would not happen under normal circumstances, but it may
happen if a malicious person gained administrative access to the VPN
gateway, or if a user was tricked into connecting to a bogus gateway.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>