Support IPv6 split tunnelling
First step is to send the "X-CSTP-Full-IPv6-Capability: true" header,
which causes the ASA to send the extra headers we need to support split
tunnelling.

Instead of receiving these headers (where a default route is implied):

 X-CSTP-Address: 2001:db8:1000:1000::1
 X-CSTP-Netmask: 2001:db8:1000:1000::1/64

Sending the aforementioned "X-CSTP-Full-IPv6-Capability: true" results
in this being received:

 X-CSTP-Address-IP6: 2001:db8:1000:1000::1/64
 X-CSTP-Split-Include-IP6: 2001:db8:1000:1001::/64
 X-CSTP-Split-Include-IP6: 2001:db8:1000:1002::/64

Because the address and netmask are being specified in the same header,
this is the reason for the slight logic change with the
"X-CSTP-Address-IP6" header parsing.

Signed-off-by: Jeremy Visser <jeremy@visser.name>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>