http: Don't retry on user cancellation
We aren't currently checking to see if do_https_request() failed due to
an actual network error or a user cancellation.  So if the library tries
to connect to an unreachable host, and the user hits Cancel before the
operation times out, we will still try falling back to non-xmlpost mode:

    $ ./openconnect 1.2.3.4
    POST https://1.2.3.4/
    Attempting to connect to server 1.2.3.4:443
    ^CSocket connect cancelled
    Failed to connect to host 1.2.3.4
    Failed to open HTTPS connection to 1.2.3.4
    GET https://1.2.3.4/
    Attempting to connect to server 1.2.3.4:443
    ^CSocket connect cancelled
    Failed to connect to host 1.2.3.4
    Failed to open HTTPS connection to 1.2.3.4
    Failed to obtain WebVPN cookie

Add a new check and return the appropriate error code for "user cancelled"
to openconnect_obtain_cookie().

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>