Add basic support for opening Windows TAP driver
We open it but can't return a file descriptor for it because Windows is...
different. We can't get a file descriptor, and if we could we couldn't
select() on it. We're probably going to have to play tricks with threads
and a socketpair.

It's either that or revamp the whole event loop to cope with the events
that Windows *can* manage to deal with.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>