android: Explicitly disable symbol versioning
On current Android builds, autotools fails to pass --sysroot to one of
the link commands, causing the test to fail and symbol versioning to be
disabled:

    configure:12875: arm-linux-androideabi-gcc -shared  -fPIC -DPIC conftest.o -v -Wl,-soname -Wl,conftest -o conftest -Wl,-M -Wl,conftest.map
    [...]
    /opt/android-ndk-r9c/toolchains/arm-linux-androideabi-4.8/prebuilt/linux-x86_64/bin/../lib/gcc/arm-linux-androideabi/4.8/../../../../arm-linux-androideabi/bin/ld: error: cannot open crtbegin_so.o: No such file or directory
    [...]
    collect2: error: ld returned 1 exit status
    configure:12878: $? = 1
    configure:12885: result: no

This is somewhat fortuitous, because it wouldn't work anyway.
__gmp_binvert_limb_table is accessed from $GMP/mpn/arm/mode1o.asm using
a homebrew PIC implementation.  This causes it to generate a relocation
entry when linked into libopenconnect.so.  The Android ldso (via
soinfo_elf_lookup() in Bionic) only allows global and weak symbols to
match relocation entries, so if __gmp_binvert_limb_table gets marked as a
local symbol by our version script, the program will fail to execute:

    soinfo_link_image(linker.cpp:1673): could not load library
    "libopenconnect.so.3" needed by "./openconnect"; caused by
    soinfo_relocate(linker.cpp:1013): cannot locate symbol
    "__gmp_binvert_limb_table" referenced by "libopenconnect.so.3"...
    CANNOT LINK EXECUTABLE

With no linker script, it remains a global symbol and there is no
problem.  Manually hexediting the library to switch the binding between
STB_LOCAL and STB_GLOBAL confirms that there is an issue with this one
symbol.

When make-standalone-toolchain.sh is used to build the library, the
--sysroot problem disappears because it is set implicitly, so symbol
versioning is enabled and that reveals the relocation problem.  Thus,
symbol versioning needs to be forced off (or libgmp needs to be fixed)
before we can safely use make-standalone-toolchain.sh.

Android bug report:
https://code.google.com/p/android/issues/detail?id=66048

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>