Fix Windows error translation to make GnuTLS happy
It really wants a read/write on a non-blocking socket to return EAGAIN,
which is kind of a pain because what we actually get is WSAEINPROGRESS. And
if we translate *that* to EAGAIN then our check for a pending connect()
will break. So ditch neterrno() completely and have a special case for
the result of connect, and then openconnect__win32_neterrno() is for
GnuTLS only.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>