Move set_tun_mtu() into os_setup_tun()
There's no real reason for this to be in openconnect_setup_tun_device()
after the vpnc-script is run. That's just a historical artifact because
we *used* to attempt to do IP configuration ourselves if there was no
vpnc-script, and we used to do the MTU setup even if there *was*.

In commit ad9c6573c ("Don't attempt to configure Legacy IP address on
tun device.") it was changed to run the script, then set the MTU anyway
just in case the script didn't get it right. But the MTU could actually
have been set much sooner.

In fact it probably doesn't need to be done at all these days, since
nobody should still be using a vpnc-script that old. But it doesn't hurt.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>