Clean up search_taps() function
Firstly, make it capable of returning an intptr_t (which will be a HANDLE)
which is generated by the callback function. Or -1 on all errors.

Secondly, add an explicit break after calling the callback, so that we
aren't making assumptions about what a zero result means.

Finally, add an 'all' argument to avoid said break, so that it can still
be used for the '--list-taps' functionality as was originally envisaged.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>