Do not abort when GnuTLS reports unclean shutdown
Some servers are observed to send an HTTP 1.0 response and then just kill
the socket. This is horrid, but we need to cope. If we get the
GNUTLS_E_PREMATURE_TERMINATION "error" when trying to read an empty
HTTP body, we shouldn't abort the connection attempt. We should just
treat it as EOF.

https://bugs.launchpad.net/bugs/1225276

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>