Add RFC4226 HOTP token support
This isn't really complete since it doesn't handle the token counter
properly. It relies on being given the token counter along with the
secret key, and there's no way to save the new value when we're done.

We could perhaps add a library function to write the token counter back,
and rely on the library user to manage the file storage containing the
counter.

Or maybe we want to use libpskc and allow the PSKC file to be specified,
then we can update that file directly.

A UI tool might also want to store the PSKC data in something like the
keyring instead of a simple file, so in that case the library should
probably allow for a callback which provides the new PSKC data rather
than unconditionally writing it to a file.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>