Fix GnuTLS 2.x build
Commit c7077b96b ("Do not abort when GnuTLS reports unclean shutdown")
introduced a build failure for GnuTLS 2.x which lacks the
GNUTLS_E_PREMATURE_TERMINATION error that we check for. Instead, it
returns GNUTLS_E_UNEXPECTED_PACKET_LENGTH.

Since the offending server turned out not to be an AnyConnect server at
all in the end, gracefully handling the unclean SSL shutdown probably
isn't strictly necessary — although it remains a good idea on general
principles. But it's not *so* important that I'm going to jump through
hoops to do it even for users who are still using GnuTLS 2.x.

So just put the check inside #ifdef GNUTLS_E_PREMATURE_TERMINATION, and
if we later find an *AnyConnect* server which displays this behaviour and
requires us to cope with it, we can find a better solution. Which might
just be to deprecate GnuTLS 2.x entirely... ☺

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>